move address from common to near specific code and wrapping AccountId
